add `fire_event` option to salt state compiler
Adding  `- fire_event: True` to a state will cause an event to be
sent with the event `name` in the tag and the state return as the
event data.

Adding  `- fire_event: my/custom/event/tag` to a state will cause an
event to be sent with 'my/custom/event/tag` in the tag and the state
return as the event data.