Refactor major functions
This begins the work of refactoring Salt's crypto implementation to make it pluggable.

The general approach is somewhat similar to the pluggable transports, wherein a factory class
selects the appropriate backend.

This is just an initial draft. There are several problems here, including the need to pass in an entire opts
dictionary around where prior to this it was not necessary.